requirements.txt
Flask==3.0.0
icalendar==5.0.11
python-dateutil==2.8.2
requests==2.31.0
pytz==2023.3
python-dotenv==1.0.0
